+++ parsers/sne/unistra/cosmo_parser_h2_unistra.py
@@ -5,7 +5,7 @@
 import pandas as pd
 import logging
 
-from scripts.data_loaders import BaseParser
+from scripts.data_loaders import BaseParser, register_parser
 
 def parse_unistra_h2_style(filepath, **kwargs):
     logger = logging.getLogger()
@@ -56,3 +56,12 @@ class UniStraH2Parser(BaseParser):
 
     def parse(self, filepath, **kwargs):
         return parse_unistra_h2_style(filepath, **kwargs)
+
+
+register_parser(
+    data_type="sne",
+    source="unistra",
+    name="Unistra H2",
+    parser=UniStraH2Parser()
+)
+